+++ cpp_wrappers/cpp_subsampling/grid_subsampling/grid_subsampling.h
