+++ cpp_wrappers/cpp_utils/cloud/cloud.cpp
