+++ data_utils/kpconv_loader.py
@@ -1,0 +1,799 @@
+import numpy as np
+from jittor.dataset.dataset import Dataset
+import time
+import pickle
+from os.path import exists, join
+from datasets.ModelNet40 import ModelNet40CustomBatch
+import cpp_wrappers.cpp_subsampling.grid_subsampling as cpp_subsampling
+import cpp_wrappers.cpp_neighbors.radius_neighbors as cpp_neighbors
+
+def batch_grid_subsampling(points, batches_len, features=None, labels=None,
+                           sampleDl=0.1, max_p=0, verbose=0, random_grid_orient=True):
+    """
+    CPP wrapper for a grid subsampling (method = barycenter for points and features)
+    :param points: (N, 3) matrix of input points
+    :param features: optional (N, d) matrix of features (floating number)
+    :param labels: optional (N,) matrix of integer labels
+    :param sampleDl: parameter defining the size of grid voxels
+    :param verbose: 1 to display
+    :return: subsampled points, with features and/or labels depending of the input
+    """
+
+    R = None
+    B = len(batches_len)
+    if random_grid_orient:
+
+        ########################################################
+        # Create a random rotation matrix for each batch element
+        ########################################################
+
+        # Choose two random angles for the first vector in polar coordinates
+        theta = np.random.rand(B) * 2 * np.pi
+        phi = (np.random.rand(B) - 0.5) * np.pi
+
+        # Create the first vector in carthesian coordinates
+        u = np.vstack([np.cos(theta) * np.cos(phi), np.sin(theta) * np.cos(phi), np.sin(phi)])
+
+        # Choose a random rotation angle
+        alpha = np.random.rand(B) * 2 * np.pi
+
+        # Create the rotation matrix with this vector and angle
+        R = create_3D_rotations(u.T, alpha).astype(np.float32)
+
+        #################
+        # Apply rotations
+        #################
+
+        i0 = 0
+        points = points.copy()
+        for bi, length in enumerate(batches_len):
+            # Apply the rotation
+            points[i0:i0 + length, :] = np.sum(np.expand_dims(points[i0:i0 + length, :], 2) * R[bi], axis=1)
+            i0 += length
+
+    #######################
+    # Sunsample and realign
+    #######################
+
+    if (features is None) and (labels is None):
+        s_points, s_len = cpp_subsampling.subsample_batch(points,
+                                                          batches_len,
+                                                          sampleDl=sampleDl,
+                                                          max_p=max_p,
+                                                          verbose=verbose)
+        if random_grid_orient:
+            i0 = 0
+            for bi, length in enumerate(s_len):
+                s_points[i0:i0 + length, :] = np.sum(np.expand_dims(s_points[i0:i0 + length, :], 2) * R[bi].T, axis=1)
+                i0 += length
+        return s_points, s_len
+
+    elif (labels is None):
+        s_points, s_len, s_features = cpp_subsampling.subsample_batch(points,
+                                                                      batches_len,
+                                                                      features=features,
+                                                                      sampleDl=sampleDl,
+                                                                      max_p=max_p,
+                                                                      verbose=verbose)
+        if random_grid_orient:
+            i0 = 0
+            for bi, length in enumerate(s_len):
+                # Apply the rotation
+                s_points[i0:i0 + length, :] = np.sum(np.expand_dims(s_points[i0:i0 + length, :], 2) * R[bi].T, axis=1)
+                i0 += length
+        return s_points, s_len, s_features
+
+    elif (features is None):
+        s_points, s_len, s_labels = cpp_subsampling.subsample_batch(points,
+                                                                    batches_len,
+                                                                    classes=labels,
+                                                                    sampleDl=sampleDl,
+                                                                    max_p=max_p,
+                                                                    verbose=verbose)
+        if random_grid_orient:
+            i0 = 0
+            for bi, length in enumerate(s_len):
+                # Apply the rotation
+                s_points[i0:i0 + length, :] = np.sum(np.expand_dims(s_points[i0:i0 + length, :], 2) * R[bi].T, axis=1)
+                i0 += length
+        return s_points, s_len, s_labels
+
+    else:
+        s_points, s_len, s_features, s_labels = cpp_subsampling.subsample_batch(points,
+                                                                              batches_len,
+                                                                              features=features,
+                                                                              classes=labels,
+                                                                              sampleDl=sampleDl,
+                                                                              max_p=max_p,
+                                                                              verbose=verbose)
+        if random_grid_orient:
+            i0 = 0
+            for bi, length in enumerate(s_len):
+                # Apply the rotation
+                s_points[i0:i0 + length, :] = np.sum(np.expand_dims(s_points[i0:i0 + length, :], 2) * R[bi].T, axis=1)
+                i0 += length
+        return s_points, s_len, s_features, s_labels
+
+def batch_neighbors(queries, supports, q_batches, s_batches, radius):
+        """
+        Computes neighbors for a batch of queries and supports
+        :param queries: (N1, 3) the query points
+        :param supports: (N2, 3) the support points
+        :param q_batches: (B) the list of lengths of batch elements in queries
+        :param s_batches: (B)the list of lengths of batch elements in supports
+        :param radius: float32
+        :return: neighbors indices
+        """
+
+        return cpp_neighbors.batch_query(queries, supports, q_batches, s_batches, radius=radius)
+
+def create_3D_rotations(axis, angle):
+    """
+    Create rotation matrices from a list of axes and angles. Code from wikipedia on quaternions
+    :param axis: float32[N, 3]
+    :param angle: float32[N,]
+    :return: float32[N, 3, 3]
+    """
+
+    t1 = np.cos(angle)
+    t2 = 1 - t1
+    t3 = axis[:, 0] * axis[:, 0]
+    t6 = t2 * axis[:, 0]
+    t7 = t6 * axis[:, 1]
+    t8 = np.sin(angle)
+    t9 = t8 * axis[:, 2]
+    t11 = t6 * axis[:, 2]
+    t12 = t8 * axis[:, 1]
+    t15 = axis[:, 1] * axis[:, 1]
+    t19 = t2 * axis[:, 1] * axis[:, 2]
+    t20 = t8 * axis[:, 0]
+    t24 = axis[:, 2] * axis[:, 2]
+    R = np.stack([t1 + t2 * t3,
+                  t7 - t9,
+                  t11 + t12,
+                  t7 + t9,
+                  t1 + t2 * t15,
+                  t19 - t20,
+                  t11 - t12,
+                  t19 + t20,
+                  t1 + t2 * t24], axis=1)
+
+    return np.reshape(R, (-1, 3, 3))
+
+class KPConvLoader(Dataset):
+    def __init__(self, config, train: bool, use_potential=True, balance_labels=False, num_workers=4, shuffle=False):
+        super().__init__(num_workers=num_workers)
+        """
+        This dataset is small enough to be stored in-memory, so load all point clouds here
+        """
+
+        self.neighborhood_limits = []
+        ############
+        # Parameters
+        ############
+
+        # Dict from labels to names
+        self.label_to_names = {0: 'airplane',
+                               1: 'bathtub',
+                               2: 'bed',
+                               3: 'bench',
+                               4: 'bookshelf',
+                               5: 'bottle',
+                               6: 'bowl',
+                               7: 'car',
+                               8: 'chair',
+                               9: 'cone',
+                               10: 'cup',
+                               11: 'curtain',
+                               12: 'desk',
+                               13: 'door',
+                               14: 'dresser',
+                               15: 'flower_pot',
+                               16: 'glass_box',
+                               17: 'guitar',
+                               18: 'keyboard',
+                               19: 'lamp',
+                               20: 'laptop',
+                               21: 'mantel',
+                               22: 'monitor',
+                               23: 'night_stand',
+                               24: 'person',
+                               25: 'piano',
+                               26: 'plant',
+                               27: 'radio',
+                               28: 'range_hood',
+                               29: 'sink',
+                               30: 'sofa',
+                               31: 'stairs',
+                               32: 'stool',
+                               33: 'table',
+                               34: 'tent',
+                               35: 'toilet',
+                               36: 'tv_stand',
+                               37: 'vase',
+                               38: 'wardrobe',
+                               39: 'xbox'}
+
+        # Initialize a bunch of variables concerning class labels
+        self.init_labels()
+
+        # List of classes ignored during training (can be empty)
+        self.ignored_labels = np.array([])
+
+        # Dataset folder
+        self.path = '../Data/ModelNet40'
+
+        # Type of task conducted on this dataset
+        self.dataset_task = 'classification'
+
+        # Update number of class and data task in configuration
+        config.num_classes = self.num_classes
+        config.dataset_task = self.dataset_task
+
+        # Parameters from config
+        self.config = config
+
+        # Training or test set
+        self.train = train
+
+        # Number of models and models used per epoch
+        if self.train:
+            self.num_models = 9843
+            if config.epoch_steps and config.epoch_steps * config.batch_num < self.num_models:
+                self.epoch_n = config.epoch_steps * config.batch_num
+            else:
+                self.epoch_n = self.num_models
+            self.epoch_n = self.num_models # 强制加上的
+            print("trainer, epoch_n = ", self.epoch_n)
+        else:
+            self.num_models = 2468
+            self.epoch_n = min(self.num_models, config.validation_size * config.batch_num)
+            print("val, epoch_n = ", self.epoch_n)
+
+        #############
+        # Load models
+        #############
+
+        if 0 < self.config.first_subsampling_dl <= 0.01:
+            raise ValueError('subsampling_parameter too low (should be over 1 cm')
+
+        self.input_points, self.input_normals, self.input_labels = self.load_subsampled_clouds()
+
+        # Does the sampler use potential for regular sampling
+        self.use_potential = use_potential
+
+        # Should be balance the classes when sampling
+        self.balance_labels = balance_labels
+
+        # Create potentials
+        if self.use_potential:
+            self.potentials = np.random.rand(len(self.input_labels)) * 0.1 + 0.1
+        else:
+            self.potentials = None
+
+        # Initialize value for batch limit (max number of points per batch).
+        self.batch_limit = 10000
+        self.batch_list = []
+        self.calibration()
+        self.prepare_batch_indices()
+        self.set_attrs(
+            batch_size=1,
+            total_len=len(self.batch_list),
+            shuffle=shuffle
+        )
+
+    def __getitem__(self, index):
+        ###################
+        # Gather batch data
+        ###################
+        idx_list = self.batch_list[index]
+
+        if isinstance(idx_list, int):
+            idx_list = [idx_list]
+
+        tp_list = []
+        tn_list = []
+        tl_list = []
+        ti_list = []
+        s_list = []
+        R_list = []
+
+        for p_i in idx_list:
+            # Get points and labels
+            points = self.input_points[p_i].astype(np.float32)
+            normals = self.input_normals[p_i].astype(np.float32)
+            label = self.label_to_idx[self.input_labels[p_i]]
+
+            # Data augmentation
+            points, normals, scale, R = self.augmentation_transform(points, normals)
+
+            # Stack batch
+            tp_list += [points]
+            tn_list += [normals]
+            tl_list += [label]
+            ti_list += [p_i]
+            s_list += [scale]
+            R_list += [R]
+
+        ###################
+        # Concatenate batch
+        ###################
+
+        #show_ModelNet_examples(tp_list, cloud_normals=tn_list)
+
+        stacked_points = np.concatenate(tp_list, axis=0)
+        stacked_normals = np.concatenate(tn_list, axis=0)
+        labels = np.array(tl_list, dtype=np.int64)
+        model_inds = np.array(ti_list, dtype=np.int32)
+        stack_lengths = np.array([tp.shape[0] for tp in tp_list], dtype=np.int32)
+        scales = np.array(s_list, dtype=np.float32)
+        rots = np.stack(R_list, axis=0)
+
+        # Input features
+        stacked_features = np.ones_like(stacked_points[:, :1], dtype=np.float32)
+        if self.config.in_features_dim == 1:
+            pass
+        elif self.config.in_features_dim == 4:
+            stacked_features = np.hstack((stacked_features, stacked_normals))
+        else:
+            raise ValueError('Only accepted input dimensions are 1, 4 and 7 (without and with XYZ)')
+
+        #######################
+        # Create network inputs
+        #######################
+        #
+        #   Points, neighbors, pooling indices for each layers
+        #
+
+        # Get the whole input list
+        input_list = self.classification_inputs(stacked_points,
+                                                stacked_features,
+                                                labels,
+                                                stack_lengths)
+
+        # Add scale and rotation for testing
+        input_list += [scales, rots, model_inds]
+
+        return input_list
+
+    def augmentation_transform(self, points, normals=None, verbose=False):
+        """Implementation of an augmentation transform for point clouds."""
+
+        ##########
+        # Rotation
+        ##########
+
+        # Initialize rotation matrix
+        R = np.eye(points.shape[1])
+
+        if points.shape[1] == 3:
+            if self.config.augment_rotation == 'vertical':
+
+                # Create random rotations
+                theta = np.random.rand() * 2 * np.pi
+                c, s = np.cos(theta), np.sin(theta)
+                R = np.array([[c, -s, 0], [s, c, 0], [0, 0, 1]], dtype=np.float32)
+
+            elif self.config.augment_rotation == 'all':
+
+                # Choose two random angles for the first vector in polar coordinates
+                theta = np.random.rand() * 2 * np.pi
+                phi = (np.random.rand() - 0.5) * np.pi
+
+                # Create the first vector in carthesian coordinates
+                u = np.array([np.cos(theta) * np.cos(phi), np.sin(theta) * np.cos(phi), np.sin(phi)])
+
+                # Choose a random rotation angle
+                alpha = np.random.rand() * 2 * np.pi
+
+                # Create the rotation matrix with this vector and angle
+                R = create_3D_rotations(np.reshape(u, (1, -1)), np.reshape(alpha, (1, -1)))[0]
+
+        R = R.astype(np.float32)
+
+        #######
+        # Scale
+        #######
+
+        # Choose random scales for each example
+        min_s = self.config.augment_scale_min
+        max_s = self.config.augment_scale_max
+        if self.config.augment_scale_anisotropic:
+            scale = np.random.rand(points.shape[1]) * (max_s - min_s) + min_s
+        else:
+            scale = np.random.rand() * (max_s - min_s) + min_s
+
+        # Add random symmetries to the scale factor
+        symmetries = np.array(self.config.augment_symmetries).astype(np.int32)
+        symmetries *= np.random.randint(2, size=points.shape[1])
+        scale = (scale * (1 - symmetries * 2)).astype(np.float32)
+
+        #######
+        # Noise
+        #######
+
+        noise = (np.random.randn(points.shape[0], points.shape[1]) * self.config.augment_noise).astype(np.float32)
+
+        ##################
+        # Apply transforms
+        ##################
+
+        # Do not use np.dot because it is multi-threaded
+        #augmented_points = np.dot(points, R) * scale + noise
+        augmented_points = np.sum(np.expand_dims(points, 2) * R, axis=1) * scale + noise
+
+
+        if normals is None:
+            return augmented_points, scale, R
+        else:
+            # Anisotropic scale of the normals thanks to cross product formula
+            normal_scale = scale[[1, 2, 0]] * scale[[2, 0, 1]]
+            augmented_normals = np.dot(normals, R) * normal_scale
+            # Renormalise
+            augmented_normals *= 1 / (np.linalg.norm(augmented_normals, axis=1, keepdims=True) + 1e-6)
+
+            return augmented_points, augmented_normals, scale, R
+
+    def classification_inputs(self,
+                              stacked_points,
+                              stacked_features,
+                              labels,
+                              stack_lengths):
+
+        # Starting radius of convolutions
+        r_normal = self.config.first_subsampling_dl * self.config.conv_radius
+
+        # Starting layer
+        layer_blocks = []
+
+        # Lists of inputs
+        input_points = []
+        input_neighbors = []
+        input_pools = []
+        input_stack_lengths = []
+        deform_layers = []
+
+        ######################
+        # Loop over the blocks
+        ######################
+
+        arch = self.config.architecture
+
+        for block_i, block in enumerate(arch):
+
+            # Get all blocks of the layer
+            if not ('pool' in block or 'strided' in block or 'global' in block or 'upsample' in block):
+                layer_blocks += [block]
+                continue
+
+            # Convolution neighbors indices
+            # *****************************
+
+            deform_layer = False
+            if layer_blocks:
+                # Convolutions are done in this layer, compute the neighbors with the good radius
+                if np.any(['deformable' in blck for blck in layer_blocks]):
+                    r = r_normal * self.config.deform_radius / self.config.conv_radius
+                    deform_layer = True
+                else:
+                    r = r_normal
+                conv_i = batch_neighbors(stacked_points, stacked_points, stack_lengths, stack_lengths, r)
+
+            else:
+                # This layer only perform pooling, no neighbors required
+                conv_i = np.zeros((0, 1), dtype=np.int32)
+
+            # Pooling neighbors indices
+            # *************************
+
+            # If end of layer is a pooling operation
+            if 'pool' in block or 'strided' in block:
+
+                # New subsampling length
+                dl = 2 * r_normal / self.config.conv_radius
+
+                # Subsampled points
+                pool_p, pool_b = batch_grid_subsampling(stacked_points, stack_lengths, sampleDl=dl)
+
+                # Radius of pooled neighbors
+                if 'deformable' in block:
+                    r = r_normal * self.config.deform_radius / self.config.conv_radius
+                    deform_layer = True
+                else:
+                    r = r_normal
+
+                # Subsample indices
+                pool_i = batch_neighbors(pool_p, stacked_points, pool_b, stack_lengths, r)
+
+            else:
+                # No pooling in the end of this layer, no pooling indices required
+                pool_i = np.zeros((0, 1), dtype=np.int32)
+                pool_p = np.zeros((0, 1), dtype=np.float32)
+                pool_b = np.zeros((0,), dtype=np.int32)
+
+            # Reduce size of neighbors matrices by eliminating furthest point
+            conv_i = self.big_neighborhood_filter(conv_i, len(input_points))
+            pool_i = self.big_neighborhood_filter(pool_i, len(input_points))
+
+            # Updating input lists
+            input_points += [stacked_points]
+            input_neighbors += [conv_i.astype(np.int64)]
+            input_pools += [pool_i.astype(np.int64)]
+            input_stack_lengths += [stack_lengths]
+            deform_layers += [deform_layer]
+
+            # New points for next layer
+            stacked_points = pool_p
+            stack_lengths = pool_b
+
+            # Update radius and reset blocks
+            r_normal *= 2
+            layer_blocks = []
+
+            # Stop when meeting a global pooling or upsampling
+            if 'global' in block or 'upsample' in block:
+                break
+
+        ###############
+        # Return inputs
+        ###############
+
+        # Save deform layers
+
+        # list of network inputs
+        li = input_points + input_neighbors + input_pools + input_stack_lengths
+        li += [stacked_features, labels]
+
+        return li
+
+    def init_labels(self):
+
+        # Initialize all label parameters given the label_to_names dict
+        self.num_classes = len(self.label_to_names)
+        self.label_values = np.sort([k for k, v in self.label_to_names.items()])
+        self.label_names = [self.label_to_names[k] for k in self.label_values]
+        self.label_to_idx = {l: i for i, l in enumerate(self.label_values)}
+        self.name_to_label = {v: k for k, v in self.label_to_names.items()}
+
+    def load_subsampled_clouds(self, orient_correction=True):
+
+        # Restart timer
+        t0 = time.time()
+
+        # Load wanted points if possible
+        if self.train:
+            split ='training'
+        else:
+            split = 'test'
+
+        print('\nLoading {:s} points subsampled at {:.3f}'.format(split, self.config.first_subsampling_dl))
+        filename = join(self.path, '{:s}_{:.3f}_record.pkl'.format(split, self.config.first_subsampling_dl))
+        print("trying to load data from ", filename)
+        if exists(filename):
+            with open(filename, 'rb') as file:
+                input_points, input_normals, input_labels = pickle.load(file)
+        # Else compute them from original points
+        else:
+            # Collect training file names
+            print("subsampled_clouds not loaded, exit!")
+            exit(0)
+
+        lengths = [p.shape[0] for p in input_points]
+        sizes = [l * 4 * 6 for l in lengths]
+        print('{:.1f} MB loaded in {:.1f}s'.format(np.sum(sizes) * 1e-6, time.time() - t0))
+
+        if orient_correction:
+            input_points = [pp[:, [0, 2, 1]] for pp in input_points]
+            input_normals = [nn[:, [0, 2, 1]] for nn in input_normals]
+
+        return input_points, input_normals, input_labels
+
+    def calibration(self, verbose=False):
+        """
+        Method performing batch and neighbors calibration.
+            Batch calibration: Set "batch_limit" (the maximum number of points allowed in every batch) so that the
+                               average batch size (number of stacked pointclouds) is the one asked.
+        Neighbors calibration: Set the "neighborhood_limits" (the maximum number of neighbors allowed in convolutions)
+                               so that 90% of the neighborhoods remain untouched. There is a limit for each layer.
+        """
+        
+        class bcolors:
+            HEADER = '\033[95m'
+            OKBLUE = '\033[94m'
+            OKGREEN = '\033[92m'
+            WARNING = '\033[93m'
+            FAIL = '\033[91m'
+            ENDC = '\033[0m'
+            BOLD = '\033[1m'
+            UNDERLINE = '\033[4m'
+        ##############################
+        # Previously saved calibration
+        ##############################
+
+        print('\nStarting Calibration (use verbose=True for more details)')
+        t0 = time.time()
+
+        redo = False
+
+        # Batch limit
+        # ***********
+
+        # Load batch_limit dictionary
+        batch_lim_file = join(self.path, 'batch_limits.pkl')
+        if exists(batch_lim_file):
+            with open(batch_lim_file, 'rb') as file:
+                batch_lim_dict = pickle.load(file)
+        else:
+            batch_lim_dict = {}
+
+        # Check if the batch limit associated with current parameters exists
+        key = '{:.3f}_{:d}'.format(self.config.first_subsampling_dl,
+                                   self.config.batch_num)
+        if key in batch_lim_dict:
+            self.batch_limit = batch_lim_dict[key]
+            # self.batch_limit //= 4
+            print("batch_limit: ", self.batch_limit)
+        else:
+            redo = True
+
+        if verbose:
+            print('\nPrevious calibration found:')
+            print('Check batch limit dictionary')
+            if key in batch_lim_dict:
+                color = bcolors.OKGREEN
+                v = str(int(batch_lim_dict[key]))
+            else:
+                color = bcolors.FAIL
+                v = '?'
+            print('{:}\"{:s}\": {:s}{:}'.format(color, key, v, bcolors.ENDC))
+
+        # Neighbors limit
+        # ***************
+
+        # Load neighb_limits dictionary
+        neighb_lim_file = join(self.path, 'neighbors_limits.pkl')
+        if exists(neighb_lim_file):
+            with open(neighb_lim_file, 'rb') as file:
+                neighb_lim_dict = pickle.load(file)
+        else:
+            neighb_lim_dict = {}
+
+        # Check if the limit associated with current parameters exists (for each layer)
+        neighb_limits = []
+        for layer_ind in range(self.config.num_layers):
+
+            dl = self.config.first_subsampling_dl * (2**layer_ind)
+            if self.config.deform_layers[layer_ind]:
+                r = dl * self.config.deform_radius
+            else:
+                r = dl * self.config.conv_radius
+
+            key = '{:.3f}_{:.3f}'.format(dl, r)
+            if key in neighb_lim_dict:
+                neighb_limits += [neighb_lim_dict[key]]
+
+        if len(neighb_limits) == self.config.num_layers:
+            self.neighborhood_limits = neighb_limits
+        else:
+            redo = True
+
+        if verbose:
+            print('Check neighbors limit dictionary')
+            for layer_ind in range(self.config.num_layers):
+                dl = self.config.first_subsampling_dl * (2**layer_ind)
+                if self.config.deform_layers[layer_ind]:
+                    r = dl * self.config.deform_radius
+                else:
+                    r = dl * self.config.conv_radius
+                key = '{:.3f}_{:.3f}'.format(dl, r)
+
+                if key in neighb_lim_dict:
+                    color = bcolors.OKGREEN
+                    v = str(neighb_lim_dict[key])
+                else:
+                    color = bcolors.FAIL
+                    v = '?'
+                print('{:}\"{:s}\": {:s}{:}'.format(color, key, v, bcolors.ENDC))
+
+        if redo:
+            print("Redo needed, exit!")
+            exit(0)
+
+
+        print('Calibration done in {:.1f}s\n'.format(time.time() - t0))
+        return 
+
+    def prepare_batch_indices(self):
+        """
+        Yield next batch indices here
+        """
+        self.batch_list = []
+        ##########################################
+        # Initialize the list of generated indices
+        ##########################################
+
+        if self.use_potential:
+            if self.balance_labels:
+                gen_indices = []
+                pick_n = self.epoch_n // self.num_classes + 1
+                for i, l in enumerate(self.label_values):
+
+                    # Get the potentials of the objects of this class
+                    label_inds = np.where(np.equal(self.input_labels, l))[0]
+                    class_potentials = self.potentials[label_inds]
+
+                    # Get the indices to generate thanks to potentials
+                    if pick_n < class_potentials.shape[0]:
+                        pick_indices = np.argpartition(class_potentials, pick_n)[:pick_n]
+                    else:
+                        pick_indices = np.random.permutation(class_potentials.shape[0])
+                    class_indices = label_inds[pick_indices]
+                    gen_indices.append(class_indices)
+
+                # Stack the chosen indices of all classes
+                gen_indices = np.random.permutation(np.hstack(gen_indices))
+
+            else:
+
+                # Get indices with the minimum potential
+                if self.epoch_n < self.potentials.shape[0]:
+                    gen_indices = np.argpartition(self.potentials, self.epoch_n)[:self.epoch_n]
+                else:
+                    gen_indices = np.random.permutation(self.potentials.shape[0])
+                gen_indices = np.random.permutation(gen_indices)
+
+            # Update potentials (Change the order for the next epoch)
+            self.potentials[gen_indices] = np.ceil(self.potentials[gen_indices])
+            self.potentials[gen_indices] += np.random.rand(gen_indices.shape[0]) * 0.1 + 0.1
+
+        else:
+            if self.balance_labels:
+                pick_n = self.epoch_n // self.num_classes + 1
+                gen_indices = []
+                for l in self.label_values:
+                    label_inds = np.where(np.equal(self.input_labels, l))[0]
+                    rand_inds = np.random.choice(label_inds, size=pick_n, replace=True)
+                    gen_indices += [rand_inds]
+                gen_indices = np.random.permutation(np.hstack(gen_indices))
+            else:
+                gen_indices = np.random.permutation(self.num_models)[:self.epoch_n]
+
+        ################
+        # Generator loop
+        ################
+
+        # Initialize concatenation lists
+        ti_list = []
+        batch_n = 0
+
+        # Generator loop
+        for p_i in gen_indices:
+        # for s, x in enumerate(self.input_points):
+            # Size of picked cloud
+            n = self.input_points[p_i].shape[0]
+
+            # In case batch is full, yield it and reset it
+            if batch_n + n > self.batch_limit and batch_n > 0:
+                self.batch_list.append(np.array(ti_list, dtype=np.int32))
+                ti_list = []
+                batch_n = 0
+
+            # Add data to current batch
+            ti_list += [p_i]
+
+            # Update batch size
+            batch_n += n
+
+        self.batch_list.append(np.array(ti_list, dtype=np.int32))
+
+    def big_neighborhood_filter(self, neighbors, layer):
+        """
+        Filter neighborhoods with max number of neighbors. Limit is set to keep XX% of the neighborhoods untouched.
+        Limit is computed at initialization
+        """
+
+        # crop neighbors matrix
+        if len(self.neighborhood_limits) > 0:
+            return neighbors[:, :self.neighborhood_limits[layer]]
+        else:
+            return neighbors
