+++ misc/ops.py
