+++ networks/cls/kernels/dispositions/k_015_center_3D.ply
